+++ demisto_sdk/commands/common/content/objects/pack_objects/pack.py
@@ -9,6 +9,10 @@ from demisto_sdk.commands.common.constants import (CLASSIFIERS_DIR,
                                                    CONNECTIONS_DIR,
                                                    DASHBOARDS_DIR,
                                                    DOC_FILES_DIR,
+                                                   GENERIC_DEFINITIONS_DIR,
+                                                   GENERIC_FIELDS_DIR,
+                                                   GENERIC_MODULES_DIR,
+                                                   GENERIC_TYPES_DIR,
                                                    INCIDENT_FIELDS_DIR,
                                                    INCIDENT_TYPES_DIR,
                                                    INDICATOR_FIELDS_DIR,
@@ -24,10 +28,11 @@ from demisto_sdk.commands.common.constants import (CLASSIFIERS_DIR,
                                                    TOOLS_DIR, WIDGETS_DIR)
 from demisto_sdk.commands.common.content.objects.pack_objects import (
     AgentTool, AuthorImage, Classifier, ClassifierMapper, Connection,
-    Contributors, Dashboard, DocFile, IncidentField, IncidentType,
-    IndicatorField, IndicatorType, Integration, LayoutObject, OldClassifier,
-    PackIgnore, PackMetaData, Playbook, PreProcessRules, Readme, ReleaseNote,
-    Report, Script, SecretIgnore, Widget)
+    Contributors, Dashboard, DocFile, GenericDefinition, GenericField,
+    GenericModule, GenericType, IncidentField, IncidentType, IndicatorField,
+    IndicatorType, Integration, LayoutObject, OldClassifier, PackIgnore,
+    PackMetaData, Playbook, PreProcessRules, Readme, ReleaseNote, Report,
+    Script, SecretIgnore, Widget)
 from demisto_sdk.commands.common.content.objects_factory import \
     path_to_pack_object
 from demisto_sdk.commands.test_content import tools
@@ -159,6 +164,26 @@ class Pack:
     def release_notes(self) -> Iterator[ReleaseNote]:
         return self._content_files_list_generator_factory(dir_name=RELEASE_NOTES_DIR,
                                                           suffix="md")
+
+    @property
+    def generic_definitions(self) -> Iterator[GenericDefinition]:
+        return self._content_files_list_generator_factory(dir_name=GENERIC_DEFINITIONS_DIR,
+                                                          suffix="json")
+
+    @property
+    def generic_modules(self) -> Iterator[GenericModule]:
+        return self._content_files_list_generator_factory(dir_name=GENERIC_MODULES_DIR,
+                                                          suffix="json")
+
+    @property
+    def generic_types(self) -> Iterator[GenericType]:
+        return self._content_files_list_generator_factory(dir_name=GENERIC_TYPES_DIR,
+                                                          suffix="json")
+
+    @property
+    def generic_fields(self) -> Iterator[GenericField]:
+        return self._content_files_list_generator_factory(dir_name=GENERIC_FIELDS_DIR,
+                                                          suffix="json")
 
     @property
     def tools(self) -> Iterator[AgentTool]: