+++ demisto_sdk/commands/common/content/objects_factory.py
@@ -5,7 +5,8 @@ from wcmatch.pathlib import Path
 from demisto_sdk.commands.common.constants import OLD_INDICATOR_TYPE, FileType
 from demisto_sdk.commands.common.content.objects.pack_objects import (
     AgentTool, AuthorImage, ChangeLog, Classifier, ClassifierMapper,
-    Connection, Contributors, Dashboard, DocFile, IncidentField, IncidentType,
+    Connection, Contributors, Dashboard, DocFile, GenericDefinition,
+    GenericField, GenericModule, GenericType, IncidentField, IncidentType,
     IndicatorField, IndicatorType, Integration, Layout, LayoutsContainer,
     OldClassifier, OldIndicatorType, PackIgnore, PackMetaData, Playbook,
     PreProcessRules, Readme, ReleaseNote, Report, Script, SecretIgnore, Widget)
@@ -45,6 +46,10 @@ TYPE_CONVERSION_BY_FileType = {
     FileType.POWERSHELL_FILE: '',
     FileType.PYTHON_FILE: '',
     FileType.CONTRIBUTORS: Contributors,
+    FileType.GENERIC_TYPE: GenericType,
+    FileType.GENERIC_FIELD: GenericField,
+    FileType.GENERIC_MODULE: GenericModule,
+    FileType.GENERIC_DEFINITION: GenericDefinition
 }
 
 TYPE_CONVERSION_BY_FILE_NAME = {