+++ demisto_sdk/commands/format/update_incidenttype.py
@@ -2,8 +2,6 @@ from typing import Tuple
 
 import click
 
-from demisto_sdk.commands.common.hook_validations.incident_type import \
-    IncidentTypeValidator
 from demisto_sdk.commands.format.format_constants import (ERROR_RETURN_CODE,
                                                           SKIP_RETURN_CODE,
                                                           SUCCESS_RETURN_CODE)
@@ -32,7 +30,7 @@ class IncidentTypesJSONFormat(BaseUpdateJSON):
 
     def run_format(self) -> int:
         try:
-            click.secho(f'\n======= Updating file: {self.source_file} =======', fg='white')
+            click.secho(f'\n================= Updating file {self.source_file} =================', fg='bright_blue')
             super().update_json()
             self.format_auto_extract_mode()
             self.set_default_values_as_needed()
@@ -79,4 +77,4 @@ class IncidentTypesJSONFormat(BaseUpdateJSON):
         if format_res:
             return format_res, SKIP_RETURN_CODE
         else:
-            return format_res, self.initiate_file_validator(IncidentTypeValidator)
+            return format_res, self.initiate_file_validator()