+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+* Fixed an issue where the **format** command has incorrectly recognized on which files to run when running using git.
+
+# 1.4.6
 * Fixed an issue where **validate** suggests, with no reason, running **format** on missing mandatory keys in yml file.
 * Skipped existence of TestPlaybook check on community and contribution integrations.
 * Fixed an issue where pre-commit didn't run on the demisto_sdk/commands folder.
@@ -9,6 +12,23 @@
 * Added to **validate** command a validation that a content entity or a pack name does not contain the words "partner" and "community".
 * Fixed an issue where **update-release-notes** ignores *--text* flag while using *-f*
 * Fixed the outputs validations in **validate** so enrichment commands will not be checked to have DBotScore outputs.
+* Added a new validation to require the dockerimage key to exist in an integration and script yml files.
+* Enhanced the **generate-test-playbook** command to use only integration tested on commands, rather than (possibly) other integrations implementing them.
+* Expanded unify command to support GenericModules - Unifies a GenericModule object with its Dashboards.
+* Added validators for generic objects:
+  - Generic Field validator - verify that the 'fromVersion' field is above 6.5.0, 'group' field equals 4 and 'id' field starts with the prefix 'generic_'.
+  - Generic Type validator - verify that the 'fromVersion' field is above 6.5.0
+  - Generic Module validator - verify that the 'fromVersion' field is above 6.5.0
+  - Generic Definition validator - verify that the 'fromVersion' field is above 6.5.0
+ * Expanded Format command to support Generic Objects - Fixes generic objects according to their validations.
+* Fixed an issue where the **update-release-notes** command did not handle ApiModules properly.
+* Added option to enter a dictionary or json of format `[{field_name:description}]` in the **json-to-outputs** command,
+  with the `-d` flag.
+* Improved the outputs for the **format** command.
+* Fixed an issue where the validations performed after the **format** command were inconsistent with **validate**.
+* Added to the **validate** command a validation for the author image.
+* Updated the **create-content-artifacts** command to support generic modules, definitions, fields and types.
+* Added an option to ignore errors for file paths and not only file name in .pack-ignore file.
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.
@@ -23,7 +43,6 @@
 * Fixed an issue where **openapi-codegen** tried to extract reference example outputs, leading to an exception.
 * Added an option to ignore secrets automatically when using the **init** command to create a pack.
 * Added a tool that gives the ability to temporarily suppress console output.
-* Expanded unify command to support GenericModules - Unifies a GenericModule object with its Dashboards.
 
 # 1.4.4
 * When formatting incident types with Auto-Extract rules and without mode field, the **format** command will now add the user selected mode.